Move diagnostic helper functions to helper.py
