+++ diagnostics.py
@@ -112,31 +112,6 @@ def diagnose_char(characteristic, probe_area, ion_type, bimaxwellian):
     return diagnostics
 
 
-def unpack_bimaxwellian(diagnostics):
-    t_e_cold, t_e_hot = diagnostics['T_e']
-    hot_frac = diagnostics['hot_fraction']
-    t_e_avg = reduce_bimaxwellian_temperature([t_e_cold, t_e_hot], hot_frac)
-    return diagnostics.update({'T_e_cold': t_e_cold,
-                               'T_e_hot': t_e_hot,
-                               'T_e_avg': t_e_avg}
-                              ).pop('T_e')
-
-
-def get_diagnostic_keys_units(probe_area=1.*u.mm**2, ion_type="He-4+", bimaxwellian=False):
-    # Perform diagnostic on some sample data to get all diagnostic names and units as dictionary of strings
-
-    bias = np.arange(-20, 20, 1) * u.V
-    current = ((bias.value / 100 + 0.2) ** 2 - 0.01) * u.A
-    chara = Characteristic(bias, current)
-    diagnostics = swept_probe_analysis(chara, probe_area, ion_type, bimaxwellian)
-    if bimaxwellian:
-        diagnostics = unpack_bimaxwellian(diagnostics)
-    keys_units = {key: str(unit_safe(value)) for key, value in diagnostics.items()}
-    keys_units.update({"n_e_cal": str(u.m ** -3)})
-    keys_units.update({"P_e": str(u.Pa)})
-    return keys_units
-
-
 def debug_char(characteristic, error_code, *pos):
     # A debug function to plot plateaus that cause errors
     if 5 < pos[-1] < 13 and 0.3 < pos[-2] / 71 < 0.7:  # max(characteristic.current) > 20 * u.mA
